ai_service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai==0.27.0
python-dotenv==0.19.0
requests==2.26.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
anthropic==0.3.0
python-jose==3.3.0
cryptography==3.4.7
gunicorn==20.1.0
PyPDF2==3.0.1
beautifulsoup4==4.9.3
lxml==4.9.0
python-dateutil==2.8.2
tqdm==4.66.1
scipy==1.11.4
joblib==1.0.1
pytest==6.2.5
black==21.7b0
flake8==3.9.2
mypy==0.910
python-json-logger==2.0.7
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0
passlib==1.7.4
aiofiles==23.2.1
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1 